fitur 2 tp jabatan belum

+++ src/main/java/com/apap/tugas1/repository/ProvinsiDb.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.apap.tugas1.model.ProvinsiModel;
 
 @Repository
-public interface ProvinsiDb extends JpaRepository<ProvinsiModel,Integer>{
+public interface ProvinsiDb extends JpaRepository<ProvinsiModel,Long>{
 
 }
