+++ src/main/java/com/apap/tugas1/service/ProvinsiServiceImpl.java
@@ -19,7 +19,7 @@ public class ProvinsiServiceImpl implements ProvinsiService{
 	private ProvinsiDb prov;
 
 	@Override
-	public Optional<ProvinsiModel> getProvinsiDetailById(int id) {
+	public Optional<ProvinsiModel> getProvinsiDetailById(long id) {
 		return prov.findById(id);
 	}
